refactor(studio): retire the vestigial routers/ package
Phases 1-2 collapsed every routers/<area>.py into its services/<area>.py
counterpart, leaving routers/ holding only the shared SSE helper and an
empty __init__.py. Move _sse.py to services/_sse.py (alongside the other
services/_*.py helpers) and delete the routers/ package. Rewire the three
sse_response imports (sessions x2, shows) to the new location and fix a
stale test comment that named the deleted routers/schedules.py path.

HTTP contract byte-identical (oracle: 79 routes), studio suite green.
